added edges to the graph

+++ graphCreator.h
@@ -2,12 +2,13 @@
 #define GRAPHCREATOR_H
 #include "graph.h"
 #include <cstdlib>
+#include <bits/stdc++.h>
 class GraphCreator{
     private:
         int numOfNodes;
-        float chanceOfEdge;
+        double chanceOfEdge;
     public:
-        GraphCreator(int numOfNodes, float chanceOfEdge);
+        GraphCreator(int numOfNodes, double chanceOfEdge);
         Graph* createGraph();
 };
 #endif
